Initial commit.
Contains all nine music tracks from the game.  Contains all of the pictures
that were included in the Creative Commons release of the Ares Data, except for
PICT id 701 and 720, whose purpose is unclear.  This specifically excludes
images which contain the Ambrosia logo.  Also included is a version of the Ares
credits modified to remove the reference to the Ambrosia logo and to add
attribution for compliance.

All included content is under the Creative Commons by-nc-sa 3.0 license.