Use white/clear fonts instead of black/white
This is the format that Antares actually uses. It formerly did a
preprocessing step at runtime to convert to white/clear, and this allows
that step to be removed. This format is harder to edit, but fonts are
hard to edit anyway, and a tool is really needed
(arescentral/antares#180).